requirements.txt
# Core dependencies
psycopg2-binary==2.9.10
spotipy==2.25.1
pandas==2.2.3
python-dotenv==1.0.0
requests==2.32.3

# Data pipeline
dbt-postgres==1.9.0
apache-airflow==2.10.5